+++ src/pages/newManager/components/ActionButtons.module.css
@@ -1,0 +1,5 @@
+.rightcol-buttons {
+  display: flex;
+  gap: 1rem;
+  flex-direction: row;
+}
