+++ src/pages/newManager/components/SubcategorySection.module.css
@@ -1,0 +1,6 @@
+.subcategory-container {
+  display: flex;
+  flex-direction: row;
+  gap: 0.25rem;
+  flex-wrap: wrap;
+}
